Cleanup code.

+++ app/src/main/java/com/marsht21/restaurantpicker/TempActivity.java
@@ -1,10 +1,5 @@
 package com.marsht21.restaurantpicker;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
@@ -14,12 +9,12 @@ import android.widget.Button;
 import android.widget.RatingBar;
 import android.widget.TextView;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.FirebaseAuth;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -28,14 +23,12 @@ public class TempActivity extends AppCompatActivity {
     private static final String TAG = TempActivity.class.getSimpleName();
     private Toolbar toolbar;
     private FirebaseFirestore db;
-    private FirebaseAuth mAuth;
     private TextView restaurantName;
     private RatingBar ratingBar;
     private RatingBar priceBar;
     private Button launchDirections;
     private String placeIdTemp;
     private String nameTemp;
-    private StringBuilder url;
     private Button launchPhone;
     private String phoneTemp;
     private String websiteTemp;
@@ -49,7 +42,6 @@ public class TempActivity extends AppCompatActivity {
         setContentView(R.layout.activity_temp);
         toolbar = findViewById(R.id.toolbar_temp);
         db = FirebaseFirestore.getInstance();
-        mAuth = FirebaseAuth.getInstance();
         restaurantName = findViewById(R.id.restaurantName);
         ratingBar = findViewById(R.id.ratingbar);
         priceBar = findViewById(R.id.pricebar);
@@ -62,55 +54,50 @@ public class TempActivity extends AppCompatActivity {
 
         db.collection("restaurants")
                 .get()
-                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                    @Override
-                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        if (task.isSuccessful()) {
-                            for (QueryDocumentSnapshot document : task.getResult()) {
-                                Float r = Float.parseFloat(document.get("rating").toString());
-                                Float p = Float.parseFloat(document.get("price level").toString());
-                                restaurantName.setText(document.get("name").toString());
-                                totalRatings.setText(document.get("total ratings").toString());
-                                ratingBar.setRating(r);
-                                priceBar.setRating(p);
-                                placeIdTemp = document.get("place id").toString();
-                                nameTemp = document.get("name").toString();
-                                phoneTemp = document.get("phone number").toString();
-                                websiteTemp = document.get("website").toString();
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+                            getFields(document);  //Gets selected data from Firestore to display
 
+                            StringBuilder url = buildDirectionsUrl();  // Builds url that opens directions to restaurant in Google maps
 
-                                StringBuilder url = buildDirectionsUrl();  // Builds url that opens directions to restaurant in Google maps
+                            launchDirections.setOnClickListener(v -> {  // Open directions in maps when button is pressed
+                                Uri uri = Uri.parse(String.valueOf(url));
+                                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                                startActivity(intent);
+                            });
 
-                                launchDirections.setOnClickListener(v -> {  // Open directions in maps when button is pressed
-                                    Uri uri = Uri.parse(String.valueOf(url));
-                                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                                    startActivity(intent);
-                                });
+                            launchPhone.setOnClickListener(v -> {  // Open phone number when button is pressed
+                                Uri uri = Uri.parse("tel:" + phoneTemp);
+                                Intent intent = new Intent(Intent.ACTION_DIAL, uri);
+                                startActivity(intent);
+                            });
 
-                                launchPhone.setOnClickListener(v -> {  // Open phone number when button is pressed
-                                    Uri uri = Uri.parse("tel:" + phoneTemp);
-                                    Intent intent = new Intent(Intent.ACTION_DIAL, uri);
-                                    startActivity(intent);
-                                });
+                            launchWebsite.setOnClickListener(v -> {  // Open phone number when button is pressed
+                                Uri uri = Uri.parse(websiteTemp);
+                                Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+                                startActivity(intent);
+                            });
 
-                                launchWebsite.setOnClickListener(v -> {  // Open phone number when button is pressed
-                                    Uri uri = Uri.parse(websiteTemp);
-                                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                                    startActivity(intent);
-                                });
-
-
-
-
-                                break;
-                            }
-                        } else {
-                            Log.d(TAG, "Error getting documents: ", task.getException());
+                            break;
                         }
+                    } else {
+                        Log.d(TAG, "Error getting documents: ", task.getException());
                     }
                 });
+    }
 
-
+    private void getFields(QueryDocumentSnapshot document) {
+        float r = Float.parseFloat(document.get("rating").toString());
+        float p = Float.parseFloat(document.get("price level").toString());
+        restaurantName.setText(document.get("name").toString());
+        totalRatings.setText(document.get("total ratings").toString());
+        ratingBar.setRating(r);
+        priceBar.setRating(p);
+        placeIdTemp = document.get("place id").toString();
+        nameTemp = document.get("name").toString();
+        phoneTemp = document.get("phone number").toString();
+        websiteTemp = document.get("website").toString();
     }
 
     @NotNull
@@ -133,14 +120,12 @@ public class TempActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                deleteRestaurantDocuments();
+        if (item.getItemId() == android.R.id.home) {
+            deleteRestaurantDocuments();
 
-                Intent intent = new Intent(TempActivity.this, HomeActivity.class);
-                startActivity(intent);
-                finish();
-                break;
+            Intent intent = new Intent(TempActivity.this, HomeActivity.class);
+            startActivity(intent);
+            finish();
         }
         return true;
     }
@@ -148,17 +133,15 @@ public class TempActivity extends AppCompatActivity {
     private void deleteRestaurantDocuments() {
         db.collection("restaurants")
                 .get()
-                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                    @Override
-                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        if (task.isSuccessful()) {
-                            for (QueryDocumentSnapshot document : task.getResult()) {
-                                Log.d(TAG, document.getId() + " => " + document.getData());
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+                            if (!document.getId().equals("restaurantID")) {
                                 db.collection("restaurants").document(document.getId()).delete();
                             }
-                        } else {
-                            Log.d(TAG, "Error deleting documents: ", task.getException());
                         }
+                    } else {
+                        Log.d(TAG, "Error deleting documents: ", task.getException());
                     }
                 });
     }
